--- src/components/Books.css
@@ -1,7 +1,0 @@
-*,
-*::before,
-*::after {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
