--- src/components/Error.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-export default function Error() {
-  return (
-    <div>
-     <h1>404</h1>
-     <h3>Not Found</h3>
-    </div>
-  )
-}
