--- src/components/NavBar.jsx
@@ -1,16 +1,0 @@
-import React from 'react'
-import { NavLink } from 'react-router-dom'
-
-export default function NavBar() {
-  return (
-    <div>
-     <h1>TaHtouHa Store</h1>
-     <nav>
-       <ul>
-         <NavLink to="/">Books</NavLink>
-         <NavLink to="/category">Categories</NavLink>
-       </ul>
-     </nav>
-    </div>
-  )
-}
