+++ src/redux/books/books.js
@@ -1,38 +1,16 @@
-import { v4 as uuidv4 } from 'uuid';
+import { createSlice } from '@reduxjs/toolkit';
 
-const initialState = [
-  {
-    id: uuidv4(),
-    title: 'A',
-    author: 'me',
+const options = {
+  name: 'books',
+  initialState: [],
+  reducers: {
+    addBook: (state, action) => {
+      state.push(action.payload);
+    },
+    removeBook: (state, action) => state.filter((book) => book.id !== action.payload.id),
   },
-  {
-    id: uuidv4(),
-    title: 'B',
-    author: 'you',
-  },
-];
-
-// actions' creator
-export const addBook = (book) => ({
-  type: 'Books/addBook',
-  payload: book,
-});
-
-export const removeBook = (book) => ({
-  type: 'Books/removeBook',
-  payload: book,
-});
-
-const bookReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case 'Books/addBook':
-      return [...state, action.payload];
-    case 'Books/removeBook':
-      return [...state.filter((book) => book.id !== action.payload.id)];
-    default:
-      return state;
-  }
 };
 
-export default bookReducer;
+const bookSlice = createSlice(options);
+export const { addBook, removeBook } = bookSlice.actions;
+export default bookSlice.reducer;
